Completed swimming mechanic: NEED TO REVIEW
Managed to work on and complete any and all of the swimming mechanics. I went through a process of trying to debug and research what function would work best because "Add Force" wasn't working. eventually I found a function called "Add Impulse" that got the reaction I wanted. After going through a process of tinkering and implementing a height raytrace to keep the object floating above the ground, I got the look I wanted.

This mechanic will need to be reviewed to see if it's intended purpose is achieved.