Fixed references in maps.
Reimported all of the maps along with the assets. This way all the mesh references are correct.

For future reference: do not un-check any boxes inside the asset migration tool when migrating assets. This will break references which can be very painful to fix by hand. It's a good idea to migrate everything into an empty project, clean up and organize the files there, and then migrate into the target project.